Responsive interface (#108)
* feature (time_zones): added global timezone of panel to events timezone

* fix (GA-2715): adds timezone for event and correct labels for transaction times

* fix (GA-2786): adds comprenhesive policies for support role

* refactor (roles): more comprehensive errors for unauthorization

* refactor (i18n): removes translations form all admin panel

* refactor (admins): renamed to users

* style (admin panel): prevents tooltips from hiding behind tables

* style (events): remove duplicated text areas from form

* style (customer portal): adds missing CSS and translations

* feature (users): adds promoter registration capabilities

* style (user): makes remember me box smaller

* fix (tests): fixes all api tests and rubocop

+++ app/controllers/admins/events/gtags_controller.rb
@@ -1,17 +1,15 @@
 # rubocop:disable Metrics/ClassLength
 class Admins::Events::GtagsController < Admins::Events::BaseController
-  before_action :set_presenter, only: [:index, :search]
-  before_action :set_gtag, only: [:show, :edit, :update, :destroy, :ban, :unban, :recalculate_balance, :solve_inconsistent]
+  before_action :set_gtag, except: [:index, :new, :create]
 
   def index
+    @gtags = @current_event.gtags.order(:tag_uid)
+    authorize @gtags
+    @gtags = @gtags.page(params[:page])
+
     respond_to do |format|
       format.html
-      format.csv do
-        gtags = Gtag.query_for_csv(@current_event)
-        redirect_to(admins_event_gtags_path(@current_event)) && return if gtags.empty?
-
-        send_data(CsvExporter.to_csv(gtags))
-      end
+      format.csv { send_data(CsvExporter.to_csv(Gtag.query_for_csv(@current_event))) }
     end
   end
 
@@ -19,19 +17,17 @@ class Admins::Events::GtagsController < Admins::Events::BaseController
     @transactions = @gtag.transactions.includes(:station).order(:gtag_counter)
   end
 
-  def search
-    render :index
-  end
-
   def new
     @gtag = Gtag.new
+    authorize @gtag
   end
 
   def create
     @gtag = Gtag.new(permitted_params)
+    authorize @gtag
     if @gtag.save
       flash[:notice] = I18n.t("alerts.created")
-      redirect_to admins_event_gtags_url
+      redirect_to admins_event_gtags_path
     else
       flash.now[:error] = @gtag.errors.full_messages.join(". ")
       render :new
@@ -41,7 +37,7 @@ class Admins::Events::GtagsController < Admins::Events::BaseController
   def update
     if @gtag.update(permitted_params)
       flash[:notice] = I18n.t("alerts.updated")
-      redirect_to admins_event_gtag_url(@current_event, @gtag)
+      redirect_to admins_event_gtag_path(@current_event, @gtag)
     else
       flash.now[:error] = @gtag.errors.full_messages.join(". ")
       render :edit
@@ -49,22 +45,25 @@ class Admins::Events::GtagsController < Admins::Events::BaseController
   end
 
   def destroy
-    if @gtag.destroy
-      flash[:notice] = I18n.t("alerts.destroyed")
-    else
-      flash[:error] = @gtag.errors.full_messages.join(". ")
+    respond_to do |format|
+      if @gtag.destroy
+        format.html { redirect_to admins_event_gtags_path, notice: 'GTag was successfully deleted.' }
+        format.json { render :show, status: :ok, location: admins_event_gtags_path }
+      else
+        redirect_to [:admins, @current_event, @gtag], error: @gtag.errors.full_messages.to_sentence
+      end
     end
-    redirect_to admins_event_gtags_url
   end
 
   def destroy_multiple
-    gtags = params[:gtags]
+    @gtags = params[:gtags]
+    authorize @gtags
     if gtags
       @current_event.gtags.where(id: gtags.keys).each do |gtag|
         flash[:error] = gtag.errors.full_messages.join(". ") unless gtag.destroy
       end
     end
-    redirect_to admins_event_gtags_url
+    redirect_to admins_event_gtags_path
   end
 
   def solve_inconsistent
@@ -79,16 +78,12 @@ class Admins::Events::GtagsController < Admins::Events::BaseController
 
   def ban
     @gtag.update!(banned: true)
-    # TODO: Refactor this when gtag blacklisting is defined
-    update_blacklist
-    redirect_to(admins_event_gtags_url)
+    redirect_to(admins_event_gtags_path)
   end
 
   def unban
     @gtag.update(banned: false)
-    # TODO: Refactor this when gtag blacklisting is defined
-    update_blacklist
-    redirect_to(admins_event_gtags_url)
+    redirect_to(admins_event_gtags_path)
   end
 
   def import
@@ -109,10 +104,8 @@ class Admins::Events::GtagsController < Admins::Events::BaseController
   end
 
   def sample_csv
-    header = %w(tag_uid format loyalty)
-    data = [%w(1218DECA31C9F92F card true),
-            %w(E6312A015028B0FB wristband false),
-            %w(A43FE1C5E9A622C2 wristband true)]
+    header = %w(tag_uid format)
+    data = [%w(1218DECA31C9F92F card), %w(E6312A015028B0FB wristband), %w(A43FE1C5E9A622C2 wristband)]
 
     csv_file = CsvExporter.sample(header, data)
     respond_to do |format|
@@ -122,40 +115,12 @@ class Admins::Events::GtagsController < Admins::Events::BaseController
 
   private
 
-  # TODO: Refactor this when gtag blacklisting is defined
-  def update_blacklist
-    atts = @current_event.device_settings
-    atts[:gtag_blacklist] = @current_event.gtags.banned.pluck(:tag_uid).join(",")
-    @current_event.update(device_settings: atts)
-  end
-
   def set_gtag
     @gtag = @current_event.gtags.find(params[:id])
-  end
-
-  def set_presenter
-    @list_model_presenter = ListModelPresenter.new(
-      model_name: "Gtag".constantize.model_name,
-      fetcher: @current_event.gtags,
-      search_query: params[:q],
-      page: params[:page],
-      include_for_all_items: [
-        customer: :tickets
-      ],
-      context: view_context
-    )
+    authorize @gtag
   end
 
   def permitted_params
-    params.require(:gtag).permit(
-      :event_id,
-      :tag_uid,
-      :format,
-      :redeemed,
-      :banned,
-      :loyalty,
-      :ticket_type_id,
-      :format
-    )
+    params.require(:gtag).permit(:event_id, :tag_uid, :format, :redeemed, :banned, :loyalty, :ticket_type_id, :format)
   end
 end
